+++ changelog.md
@@ -1,3 +1,8 @@
+#### Version 5.9.0
+* Add `showFrame`, `fameInnerPadding`, `frameOuterPadding` properties to `primitives.orgdiagram.Config` & `primitives.famdiagram.Config`
+* Modified callout annotation palacement for minimized nodes placed outside view port.
+* Fixed exception when `selectedItems` collection contains non existing items.
+* Fixed center on cursor in AutoSize mode
 #### Version 5.8.2
 * Fixed `selectionPathMode` in Family Diagram.
 #### Version 5.8.1